Renamed tests to NodeNormTesting.
