+++ website/src/views/NameResValidator.vue
@@ -3,8 +3,9 @@
   <p>
     <a href="https://github.com/TranslatorSRI/babel">Babel</a> is the program that generates the datasets of
     interrelated identifiers that underlies <a href="https://github.com/TranslatorSRI/NodeNormalization">Node Normalization</a>
-    and <a href="https://github.com/TranslatorSRI/NameResolution">Name Resolution</a>. This website is intended to provide
-    a single place where tests can be executed against multiple versions of these tools.
+    and <a href="https://github.com/TranslatorSRI/NameResolution">Name Resolution</a>.
+
+    This page will test several instances of the Name Resolver.
   </p>
 
   <b-button @click="loadGoogleSheet()">Reload</b-button>
@@ -22,8 +23,8 @@
       <tr>
         <th>Test</th>
         <th>Source</th>
-        <th v-for="endpoint in Object.keys(nodeNormEndpoints)">
-          <a target="_blank" :href="nodeNormEndpoints[endpoint] + '/docs'">{{endpoint}}</a>
+        <th v-for="endpoint in Object.keys(nameResEndpoints)">
+          <a target="_blank" :href="nameResEndpoints[endpoint] + '/docs'">{{endpoint}}</a>
         </th>
       </tr>
     </thead>
@@ -31,8 +32,8 @@
       <tr v-for="test in tests">
         <td><TextWithURLs :text="test.description" :urls="test.urls"></TextWithURLs></td>
         <td><TextWithURLs :text="test.source" :urls="{'URL': test.source_url}"></TextWithURLs></td>
-        <td v-for="endpoint in Object.keys(nodeNormEndpoints)">
-          <TestResult :test="test" :endpoint="nodeNormEndpoints[endpoint]" :description="test.description + ':' + test.source + ':' + nodeNormEndpoints[endpoint]"></TestResult>
+        <td v-for="endpoint in Object.keys(nameResEndpoints)">
+          <TestResult :test="test" :endpoint="nameResEndpoints[endpoint]" :description="test.description + ':' + test.source + ':' + nameResEndpoints[endpoint]"></TestResult>
         </td>
       </tr>
     </tbody>
@@ -44,19 +45,19 @@
 import {BTable} from "bootstrap-vue-3";
 import Papa from 'papaparse';
 import TextWithURLs from "@/components/TextWithURLs.vue";
-import { Test } from '@/models/tests';
+import { Test } from '@/models/NodeNormTesting';
 import TestResult from "@/components/TestResult.vue";
 
 export default {
   components: {TestResult, BTable, TextWithURLs},
   data () {
     return {
-      nodeNormEndpoints: {
-        "NodeNorm-RENCI-exp": "https://nodenormalization-dev.apps.renci.org/1.3",
-        "NodeNorm-RENCI-dev": "https://nodenormalization-sri.renci.org/1.3",
-        "NodeNorm-ITRB-ci": "https://nodenorm.ci.transltr.io/1.3",
-        "NodeNorm-ITRB-test": "https://nodenorm.test.transltr.io/1.3",
-        "NodeNorm-ITRB-prod": "https://nodenorm.transltr.io/1.3"
+      nameResEndpoints: {
+        "NameRes-RENCI-exp": "https://name-resolution-sri-dev.apps.renci.org",
+        "NodeNorm-RENCI-dev": "https://name-resolution-sri.apps.renci.org",
+        "NodeNorm-ITRB-ci": "https://name-lookup.ci.transltr.io",
+        "NodeNorm-ITRB-test": "https://name-lookup.test.transltr.io",
+        "NodeNorm-ITRB-prod": "https://name-lookup.transltr.io"
       },
       testData: [],
       testDataErrors: [],
